Initialize ThreadPools only if/when needed. (#110)
Calling ThreadPool will spawn the requested number of Threads
immediately. As `tasks` is imported when the `@background` decorator is
used a new ThreadPool will be created for every Django worker if a
`@background` task is defined. Altough those threads are idle they are
using still ressources.

This commit introduces a PoolRunner which does create a single static
ThreadPool only if/when a async task is requested to run.
The downside of this approach is the additional work needed for creating
the Pool when starting the first task. But as tasks should be run from
the managment command this shouldn't be a problem.